OptimalTransportDataIntegration.jl

OptimalTransportDataIntegration.Instance

OptimalTransportDataIntegration.Solution

OptimalTransportDataIntegration.average_distance_to_closest

OptimalTransportDataIntegration.compute_pred_error!

OptimalTransportDataIntegration.entropic_partial_wasserstein

OptimalTransportDataIntegration.generate_data

OptimalTransportDataIntegration.generate_xcat_ycat

OptimalTransportDataIntegration.loss_crossentropy

OptimalTransportDataIntegration.modality_cost

OptimalTransportDataIntegration.ot_joint

OptimalTransportDataIntegration.read_params

OptimalTransportDataIntegration.save_params

OptimalTransportDataIntegration.Instance — Type

struct Instance

Definition and initialization of an Instance structure

datafile : file name

distance : ∈ ( Cityblock, Euclidean, Hamming )

indXA : indexes of subjects of A with given X value

indXB : indexes of subjects of B with given X value

OptimalTransportDataIntegration.Solution — Type

mutable struct Solution

tsolve : solution time

jointYZA : joint distribution of Y and Z in A

jointYZB : joint distribution of Y and Z in B

estimatorZA : estimator of probability of Z for individuals in base A

estimatorYB : estimator of probability of Y for individuals in base B

OptimalTransportDataIntegration.average_distance_to_closest — Method

average_distance_to_closest(inst, percent_closest)

Compute the cost between pairs of outcomes as the average distance between covariations of individuals with these outcomes, but considering only the percent closest neighbors

OptimalTransportDataIntegration.compute_pred_error! — Function

compute_pred_error!(sol, inst)
compute_pred_error!(sol, inst, proba_disp)
compute_pred_error!(sol, inst, proba_disp, mis_disp)
compute_pred_error!(
    sol,
    inst,
    proba_disp,
    mis_disp,
    full_disp
)

Compute prediction errors in a solution

OptimalTransportDataIntegration.entropic_partial_wasserstein — Method

entropic_partial_wasserstein(a, b, M, reg, m=nothing, numItermax=1000, 
    stopThr=1e-100, verbose=false, log=false)

This function is a Julia translation of the function entropic_partial_wasserstein in the Python Optimal Transport package.

Solves the partial optimal transport problem and returns the OT plan

The function considers the following problem:

\[ \gamma = \mathop{\arg \min}_\gamma \quad \langle \gamma, \mathbf{M} \rangle_F + \mathrm{reg} \cdot\Omega(\gamma) s.t. \gamma \mathbf{1} &\leq \mathbf{a} \\ \gamma^T \mathbf{1} &\leq \mathbf{b} \\ \gamma &\geq 0 \\ \mathbf{1}^T \gamma^T \mathbf{1} = m &\leq \min\{\|\mathbf{a}\|_1, \|\mathbf{b}\|_1\} \\\]

where :

$\mathbf{M}$ is the metric cost matrix

$\Omega$ is the entropic regularization term, $\Omega=\sum_{i,j} \gamma_{i,j}\log(\gamma_{i,j})$

$\mathbf{a}$ and :math:\mathbf{b} are the sample weights

$m$ is the amount of mass to be transported

The formulation of the problem has been proposed in :ref:[3] <references-entropic-partial-wasserstein> (prop. 5)

Parameters

a : Array (dima,) Unnormalized histogram of dimension `dima`

b : Array (dimb,) Unnormalized histograms of dimension `dimb`

M : Array (dima, dimb) cost matrix

reg : Float Regularization term > 0

m : Float, optional Amount of mass to be transported

numItermax : Int, optional Max number of iterations

stopThr : Float, optional Stop threshold on error (>0)

verbose : Bool, optional Print information along iterations

log : Bool, optional record log if True

Returns

gamma : (dima, dimb) ndarray Optimal transportation matrix for the given parameters

log : Dict log dictionary returned only if log is True

Examples

julia> a = [.1, .2];

julia> b = [.1, .1];

julia> M = [0. 1.; 2. 3.];

julia> round.(entropic_partial_wasserstein(a, b, M, 1, m = 0.1), digits=2)
2×2 Matrix{Float64}:
 0.06  0.02
 0.01  0.0

References

[3] Benamou, J. D., Carlier, G., Cuturi, M., Nenna, L., & Peyré, G. (2015). Iterative Bregman projections for regularized transportation problems. SIAM Journal on Scientific Computing, 37(2), A1111-A1138.

OptimalTransportDataIntegration.generate_data — Method

generate_data(params)

Function to generate data where X and (Y,Z) are categoricals

the function return a Dataframe with X1, X2, X3, Y, Z and the database id.

OptimalTransportDataIntegration.generate_xcat_ycat — Method

generate_xcat_ycat(params)

Function to generate data where X and (Y,Z) are categoricals

the function return a Dataframe with X1, X2, X3, Y, Z and the database id.

OptimalTransportDataIntegration.loss_crossentropy — Method

loss_crossentropy(Y, F)

Cross entropy is typically used as a loss in multi-class classification, in which case the labels y are given in a one-hot format. dims specifies the dimension (or the dimensions) containing the class probabilities. The prediction ŷ is usually probabilities but in our case it is also one hot encoded vector.

OptimalTransportDataIntegration.modality_cost — Method

modality_cost(loss, weight)

loss: matrix of size len(weight) * len(levels)

weight: vector of weights

Returns the scalar product <loss[level,],weight>

OptimalTransportDataIntegration.ot_joint — Function

ot_joint(inst)
ot_joint(inst, maxrelax)
ot_joint(inst, maxrelax, lambda_reg)
ot_joint(inst, maxrelax, lambda_reg, percent_closest)
ot_joint(inst, maxrelax, lambda_reg, percent_closest, norme)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol
)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol,
    full_disp
)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol,
    full_disp,
    solver_disp
)

Model where we directly compute the distribution of the outcomes for each individual or for sets of indviduals that similar values of covariates

aggregate_tol: quantify how much individuals' covariates must be close for aggregation

reg_norm: norm1, norm2 or entropy depending on the type of regularization

percent_closest: percent of closest neighbors taken into consideration in regularization

lambda_reg: coefficient measuring the importance of the regularization term

full_disp: if true, write the transported value of each individual; otherwise, juste write the number of missed transports

solver_disp: if false, do not display the outputs of the solver

OptimalTransportDataIntegration.read_params — Method

read_params(jsonfile)

Read the data generation scenario from a JSON file

OptimalTransportDataIntegration.save_params — Method

save_params(jsonfile, params)

Write the data generation scenario to a JSON file