Optimal transport of the joint distribution within a data source

OptimalTransportDataIntegration.Instance — Type

struct Instance

Definition and initialization of an Instance structure

datafile : file name

distance : ∈ ( Cityblock, Euclidean, Hamming )

indXA : indexes of subjects of A with given X value

indXB : indexes of subjects of B with given X value

OptimalTransportDataIntegration.average_distance_to_closest — Method

average_distance_to_closest(inst, percent_closest)

Compute the cost between pairs of outcomes as the average distance between covariations of individuals with these outcomes, but considering only the percent closest neighbors

OptimalTransportDataIntegration.Solution — Type

mutable struct Solution

tsolve : solution time

jointYZA : joint distribution of Y and Z in A

jointYZB : joint distribution of Y and Z in B

estimatorZA : estimator of probability of Z for individuals in base A

estimatorYB : estimator of probability of Y for individuals in base B